Improve class documentation for Plugins ; Also ensure widgets are created only when requested ; Cleanup virtual/override keywords ;

+++ Plugins/Animation/src/UI/AnimationUI.cpp
@@ -111,7 +111,7 @@ void AnimationUI::on_m_saveDir_clicked() {
 }
 
 void AnimationUI::updateTime( float t ) {
-    ui->m_animationTimeDisplay->setText( QString::number( t ) );
+    ui->m_animationTimeDisplay->setText( QString::number( double( t ) ) );
 }
 
 void AnimationUI::updateFrame( int f ) {
